Customer Retention Policy
Business Challenge/Requirement
===================================

Customer retention and acquisition strategies are on top of every organization’s agenda. 
To offer better customer service and boost loyalty, a company has to invest in a state-of-the-art CRM tool. 
In pursuit of these goals, every organization implements CRM as a strategy that integrates the concepts of data mining and data warehousing. 
The data collected through the CRM helps the leadership team make actionable decisions in real time. 
It helps them build and retain long-term and profitable relationships with customers. FutureCart Inc. is a hypothetical leading retail company 
with an omnipresence in India with more than 5000 retail stores and hypermarkets across and e-commerce in the country. The company has formed a dedicated team to handle after-sales services. The team is entrusted with the responsibility to address customer complaints and delight them - and eventually increase brand loyalty

Abstract
=============================

The company has multiple contact centers across India to provide support service to their customers

 • Customers can reach out to the care team over different communication channels depending on their preference and convenience: Calls, Chat, or Email. 
 • CCR (Customer Care Representative) registers the complaint by collecting all the necessary details - - which is called a case 
• A case can have a status -- open or closed 
• Each case can belong to a category and sub-category. This category and sub-category will determine case priority. Depending on the priority key, CCR has an SLA (in hours) to close the case within the SLA hours
 • Once a case is closed, the customer is sent a survey link to rate the overall experience of interacting with the contact center representative
 • The customer can take a survey or leave it unattended. The customer can rate the experience on a scale of 1-10 on various questions
 • Survey response is captured for that particular case